Moved to new package langdetect.
