+++ spring-xd-lang-detector/src/main/java/org/springframework/xd/analytics/linguistics/langdetect/TextModel.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.analytics.linguistics;
+package org.springframework.xd.analytics.linguistics.langdetect;
 
 /**
  * Enum for the supported text language models.
