+++ spring-xd-lang-detector/src/main/resources/config/spring-module.properties
@@ -1,1 +1,1 @@
-options_class = org.springframework.xd.analytics.linguistics.LanguageDetectorOptionsMetadata
+options_class = org.springframework.xd.analytics.linguistics.langdetect.LanguageDetectorOptionsMetadata
