+++ spring-xd-lang-detector/src/test/java/org/springframework/xd/analytics/linguistics/langdetect/DeterministicLongTextLanguageDetectorTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.xd.analytics.linguistics;
+package org.springframework.xd.analytics.linguistics.langdetect;
 
 import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.*;
@@ -24,21 +24,22 @@ import org.springframework.xd.tuple.Tuple;
 /**
  * @author Thomas Darimont
  */
-public class DeterministicShortTextLanguageDetectorTests extends AbstractLanguageDetectorTests {
+public class DeterministicLongTextLanguageDetectorTests extends AbstractLanguageDetectorTests {
 
 	@Override
 	protected LanguageDetector createNewLanguageDetector() {
 
 		LanguageDetector processor = super.createNewLanguageDetector();
+		processor.setTextModel(TextModel.LONGTEXT);
 		processor.setDeterministicLanguageDetection(true);
 
 		return processor;
 	}
 
 	@Test
-	public void testProcess_simple_short_english_text() throws Exception {
+	public void testProcess_simple_long_english_text() throws Exception {
 
-		Tuple input = newTupleWithText(Texts.SHORT_ENGLISH_1);
+		Tuple input = newTupleWithText(Texts.LONG_ENGLISH_1);
 
 		Tuple output = languageDetectionProcessor.process(input);
 
@@ -46,9 +47,9 @@ public class DeterministicShortTextLanguageDetectorTests extends AbstractLanguag
 	}
 
 	@Test
-	public void testProcess_simple_short_german_text() throws Exception {
+	public void testProcess_simple_long_german_text() throws Exception {
 
-		Tuple input = newTupleWithText(Texts.SHORT_GERMAN_1);
+		Tuple input = newTupleWithText(Texts.LONG_GERMAN_1);
 
 		Tuple output = languageDetectionProcessor.process(input);
 
@@ -58,7 +59,7 @@ public class DeterministicShortTextLanguageDetectorTests extends AbstractLanguag
 	@Test
 	public void testProcess_simple_short_italian_text() throws Exception {
 
-		Tuple input = newTupleWithText(Texts.SHORT_ITALIAN_1);
+		Tuple input = newTupleWithText(Texts.LONG_ITALIAN_1);
 
 		Tuple output = languageDetectionProcessor.process(input);
 
@@ -66,14 +67,14 @@ public class DeterministicShortTextLanguageDetectorTests extends AbstractLanguag
 	}
 
 	@Test
-	public void testProcess_simple_short_multi_language_text() throws Exception {
+	public void testProcess_simple_long_multi_language_text() throws Exception {
 
-		Tuple input = newTupleWithText(Texts.SHORT_ENGLISH_FRENCH_MIX_1);
+		Tuple input = newTupleWithText(Texts.LONG_ENGLISH_1 + Texts.LONG_GERMAN_1);
 
 		Tuple output = languageDetectionProcessor.process(input);
 
 		assertThat(output.hasFieldName(predLangOutField), is(true));
-		assertThat(output.getString(predLangOutField), is("en"));
+		assertThat(output.getString(predLangOutField), is("de"));
 
 		assertThat(output.hasFieldName(predLangProbsOutputField), is(true));
 		assertThat(output.getValue(predLangProbsOutputField), is(not(nullValue())));
@@ -83,11 +84,11 @@ public class DeterministicShortTextLanguageDetectorTests extends AbstractLanguag
 	@Test
 	public void testProcess_should_return_the_same_probabilities_for_the_same_document() throws Exception {
 
-		Tuple input = newTupleWithText(Texts.SHORT_ENGLISH_2);
+		Tuple input = newTupleWithText(Texts.LONG_GERMAN_1);
 
 		Tuple firstOutput = languageDetectionProcessor.process(input);
 		Tuple secondOutput = languageDetectionProcessor.process(input);
 
-		assertThat(extractLanguageProbability(firstOutput,"en"), is(equalTo(extractLanguageProbability(secondOutput, "en"))));
+		assertThat(extractLanguageProbability(firstOutput, "de"), is(equalTo(extractLanguageProbability(secondOutput, "de"))));
 	}
 }
