+++ spring-xd-lang-detector/src/test/java/org/springframework/xd/analytics/linguistics/langdetect/Texts.java
@@ -1,4 +1,4 @@
-package org.springframework.xd.analytics.linguistics;
+package org.springframework.xd.analytics.linguistics.langdetect;
 
 /**
  * @author Thomas Darimont
